pyproject.toml
[tool.poetry]
name = "epoch8-target-cloudpubsub"
version = "0.0.1"
description = "`target-cloudpubsub` is a Singer target for Cloud PubSub, built with the Meltano Singer SDK."
readme = "README.md"
authors = ["Andrey Tatarinov <a@tatarinov.co>"]
keywords = [
    "ELT",
    "Cloud PubSub",
]
license = "Apache-2.0"
packages = [
    { include = "target_cloudpubsub" },
]

[tool.poetry.dependencies]
python = ">=3.8,<4"
singer-sdk = { version="~=0.33.0" }
fs-s3fs = { version = "~=1.1.1", optional = true }
requests = "~=2.31.0"
google-cloud-pubsub = "^2.18.4"

[tool.poetry.dev-dependencies]
pytest = ">=7.4.0"
singer-sdk = { version="~=0.33.0", extras = ["testing"] }

[tool.poetry.extras]
s3 = ["fs-s3fs"]

[tool.ruff]
ignore = [
    "ANN101",  # missing-type-self
    "ANN102",  # missing-type-cls
]
select = ["ALL"]
src = ["target_cloudpubsub"]
target-version = "py37"

[tool.ruff.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.isort]
known-first-party = ["target_cloudpubsub"]

[tool.ruff.pydocstyle]
convention = "google"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# CLI declaration
target-cloudpubsub = 'target_cloudpubsub.target:TargetCloudPubSub.cli'
